refactor(api): 지출 트렌드 분석 기능 추가 및 관련 유틸리티 정리
- ExpenseTrendService를 추가하여 지출 트렌드 분석 로직을 분리
- 기존 ExpensesService에서 트렌드 분석 관련 메서드를 ExpenseTrendService로 이동
- PeriodType enum을 활용하여 날짜 범위 계산 유틸을 정리하고, 관련 함수들을 state 디렉토리로 이동
- 지출 통계 계산 및 트렌드 분석을 위한 새로운 유틸 함수들 추가
- 코드 가독성 및 재사용성 향상
